# backend_services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  internal_handler.SidebarItem:
    properties:
      icon:
        example: layout-dashboard
        type: string
      id:
        example: dashboard
        type: string
      label:
        example: Dashboard
        type: string
      path:
        example: dashboard
        type: string
      position:
        example: top
        type: string
    type: object
  internal_handler.SidebarResponse:
    properties:
      app_slug:
        example: my-app
        type: string
      org_id:
        example: org-123
        type: string
      sections:
        items:
          $ref: '#/definitions/internal_handler.SidebarSection'
        type: array
    type: object
  internal_handler.SidebarSection:
    properties:
      collapsible:
        example: true
        type: boolean
      default_expanded:
        example: true
        type: boolean
      id:
        example: main
        type: string
      items:
        items:
          $ref: '#/definitions/internal_handler.SidebarItem'
        type: array
      label:
        example: Main
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API gateway for Eternal Orbit Labs — authentication, organizations,
    teams, and app management.
  title: EOL Gateway API
  version: "1.0"
paths:
...
  /api/apps/{slug}/orgs/{orgId}/sidebar:
    get:
      description: Returns the sidebar navigation configuration for an organization
        workspace.
      parameters:
      - description: App slug
        in: path
        name: slug
        required: true
        type: string
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SidebarResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - CookieAuth: []
      - BearerAPIKey: []
      summary: Get org sidebar
      tags:
      - Sidebar
...
  /api/auth/google:
    get:
      description: Redirects the user to Google's consent screen. Sets an oauth_state
        cookie for CSRF protection.
      responses:
        "307":
          description: Redirect to Google
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Start Google OAuth
      tags:
      - OAuth
  /api/auth/google/callback:
    get:
      description: Handles the OAuth callback from Google. Validates state, exchanges
        code for tokens, creates or links user account, sets auth cookies, and redirects
        to frontend.
      parameters:
      - description: OAuth state parameter
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "307":
          description: Redirect to frontend with error
      summary: Google OAuth callback
      tags:
      - OAuth
...
schemes:
- http
- https
securityDefinitions:
  BearerAPIKey:
    description: Personal API key — send as "Bearer eol_k1_..."
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    description: JWT access token set automatically on login/signup.
    in: cookie
    name: eol_access
    type: apiKey
swagger: "2.0"